cms_backend/requirements.txt
# Django and DRF
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.2.0

# Environment Variables
python-dotenv==1.0.0

# Development Tools
django-debug-toolbar==4.2.0

# API Documentation
drf-yasg==1.21.7

# Security
django-filter==23.5
django-cleanup==8.0.0

# Utilities
python-slugify==8.0.1
pytz==2023.3 